refactor: move ChatModerationActionMenu and ChatModerationDetailsModal components
