+++ web/components/chat/ChatModerationActionMenu/ChatModerationActionMenu.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { RecoilRoot } from 'recoil';
-import ChatModerationDetailsModal from '../components/chat/ChatModerationActionMenu/ChatModerationDetailsModal';
+import ChatModerationActionMenu from './ChatModerationActionMenu';
 
 const mocks = {
   mocks: [
@@ -65,28 +65,33 @@ const mocks = {
 };
 
 export default {
-  title: 'owncast/Chat/Moderation modal',
-  component: ChatModerationDetailsModal,
+  title: 'owncast/Chat/Moderation menu',
+  component: ChatModerationActionMenu,
   parameters: {
     fetchMock: mocks,
     docs: {
       description: {
-        component: `This should be a modal that gives the moderator more details about the user such as:
-- When the user was created
-- Other names they've used
-- If they're authenticated, and using what method (IndieAuth, FediAuth)
-`,
+        component: `This should be a popup that is activated from a user's chat message. It should have actions to:
+- Remove single message
+- Ban user completely
+- Open modal to see user details
+        `,
       },
     },
   },
-} as ComponentMeta<typeof ChatModerationDetailsModal>;
+} as ComponentMeta<typeof ChatModerationActionMenu>;
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-const Template: ComponentStory<typeof ChatModerationDetailsModal> = args => (
+const Template: ComponentStory<typeof ChatModerationActionMenu> = args => (
   <RecoilRoot>
-    <ChatModerationDetailsModal userId="testuser123" accessToken="fakeaccesstoken4839" />
+    <ChatModerationActionMenu
+      accessToken="abc123"
+      messageID="xxx"
+      userDisplayName="Fake-User"
+      userID="abc123"
+    />
   </RecoilRoot>
 );
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const Example = Template.bind({});
+export const Basic = Template.bind({});
