+++ web/components/chat/ChatModerationDetailsModal/ChatModerationDetailsModal.module.scss
